+++ tests/testthat/test-RunCohortGeneration.R
@@ -15,6 +15,25 @@ test_that("Call runCohortGeneration without connectionDetails", {
   )
 })
 
+test_that("Call runCohortGeneration with duplicate cohort IDs across cohort sets", {
+  cohortDefinitionSet <- data.frame(cohortId = c(1, 2))
+  negativeControlOutcomeCohortSet <- data.frame(
+    cohortId = c(2, 3),
+    outcomeConceptId = c(2, 3)
+  )
+
+  expect_error(
+    runCohortGeneration(
+      connectionDetails = connectionDetails,
+      cdmDatabaseSchema = "main",
+      cohortDefinitionSet = cohortDefinitionSet,
+      negativeControlOutcomeCohortSet = negativeControlOutcomeCohortSet,
+      outputFolder = outputFolder
+    ),
+    message = "(Duplicate cohort IDs found across your cohortDefinitionSet and negativeControlOutcomeCohortSet: 2)"
+  )
+})
+
 test_that("Call runCohortGeneration with incrementalFolder specified", {
   testOutputFolder <- file.path(outputFolder, "inc_folder_test")
   on.exit(unlink(testOutputFolder, recursive = TRUE))